WIP: split stale refresh implementation for LOC gate (checkpoint)
